requirements.txt
alembic==1.4.2
click==7.1.2
dnspython==1.16.0
eventlet==0.31.0
Flask==1.1.2
Flask-Login==0.5.0
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==0.4.16
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
monotonic==1.5
psycopg2==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.1
python-environ==0.4.54
python-socketio==4.6.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.3
